use forwardable to switch between observing and tailing, improve specs
